WIP Start adding the core components to the new service
