+++ letsrobotandroid-api/src/main/java/tv/letsrobot/android/api/enums/LogLevel.kt
@@ -1,0 +1,13 @@
+package tv.letsrobot.android.api.enums
+
+/**
+ * The Class's log level
+ */
+enum class LogLevel {
+    TRACE,
+    INFO,
+    DEBUG,
+    WARNING,
+    ERROR,
+    NONE
+}
